Fix increasing the root logging level by log-capture
Log-capturing used to set the root logger level to the configured capturing level, so that it is sufficiently low to capture the desired level.
However when the root logging level is already lower, this will in fact increase the level, changing the level of all other registered handlers.

This change fixes this: Not to directly set the root logging level, but instead only decrease it if necessary.